requirements.txt
Django>=3.2.4,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.8.6,<2.9.0
python-dotenv>=0.17.1,<0.18.0
flake8>=3.9.2,<3.10.0
drf-yasg>=1.20.0,<1.21.0
